feat(rive-codegen): split monolithic main.cpp into focused modules and add --lint flag
Extracts the 1800-line main.cpp into:
- types.h: all shared enums and structs (RiveFileData, LintConfig, LintViolation, etc.)
- string_utils.h: inline case-conversion and sanitization helpers
- rive_file_parser.h/cpp: file discovery, Rive runtime parsing, asset/artboard extraction
- lint.h/cpp: LintingFactory, runLint(), outputLintViolations()
- template_renderer.h/cpp: loadTemplate(), renderTemplate(), Mustache/Inja data builders
- main.cpp: lean CLI setup and top-level flow

Adds --lint flag with configurable --allowed-extensions, --max-asset-size, and
--lint-format (text/json). Can run lint-only (no -o) or lint + codegen combined.

Assisted by AI
